requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
numpy==1.9.2
Pillow==2.8.1
pymongo==3.0
scipy==0.13.0
Werkzeug==0.10.4
